+++ node/runtime/pangolin/src/lib.rs
@@ -58,8 +58,8 @@ pub use darwinia_balances::Call as BalancesCall;
 pub use darwinia_fee_market::Call as FeeMarketCall;
 pub use frame_system::Call as SystemCall;
 pub use pallet_bridge_grandpa::Call as BridgeGrandpaCall;
-pub use pallet_bridge_parachains::Call as BridgeParachainsCall;
 pub use pallet_bridge_messages::Call as BridgeMessagesCall;
+pub use pallet_bridge_parachains::Call as BridgeParachainsCall;
 pub use pallet_sudo::Call as SudoCall;
 
 // --- crates.io ---
@@ -258,14 +258,13 @@ frame_support::construct_runtime! {
 		BridgePangoroGrandpa: pallet_bridge_grandpa::<Instance1>::{Pallet, Call, Storage} = 45,
 		BridgePangoroMessages: pallet_bridge_messages::<Instance1>::{Pallet, Call, Storage, Event<T>} = 43,
 
+		BridgeRococoGrandpa: pallet_bridge_grandpa::<Instance2>::{Pallet, Call, Storage} = 60,
+		BridgeRococoParachains: pallet_bridge_parachains::<Instance1>::{Pallet, Call, Storage} = 61,
+
 		FeeMarket: darwinia_fee_market::{Pallet, Call, Storage, Event<T>} = 53,
 		TransactionPause: module_transaction_pause::{Pallet, Call, Storage, Event<T>} = 54,
 
 		Substrate2SubstrateIssuing: from_substrate_issuing::{Pallet, Call, Storage, Config, Event<T>} = 49,
-
-		BridgeRococoGrandpa: pallet_bridge_grandpa::<Instance2>::{Pallet, Call, Storage} = 60,
-
-		BridgeRococoParachains: pallet_bridge_parachains::{Pallet, Call, Storage} = 61,
 
 		// BSC: darwinia_bridge_bsc::{Pallet, Call, Storage, Config} = 46,
 	}
